requirements.txt
absl-py==2.3.0
accelerate==1.6.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.11.1
asttokens==2.4.1
attrs==25.1.0
autograd==1.7.0
azure-core==1.39.0
azure-storage-blob==12.28.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
# contourpy==1.3.1
cryptography==46.0.7
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
dask==2025.1.0
dataframe_image==0.2.7
datasets==3.5.0
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
Django==6.0.3
docker-pycreds==0.4.0
et_xmlfile==2.0.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.56.0
Fraction==2.2.0
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grobid-client-python==0.0.9
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
# imbalanced-learn==0.13.0
# imblearn==0.0
immutabledict==4.2.1
ipykernel==6.29.5
ipython==8.27.0
isodate==0.7.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
KDEpy==1.1.12
kiwisolver==1.4.8
langdetect==1.0.9
locket==1.0.0
# lxml==5.3.0
MarkupSafe==3.0.2
# matplotlib==3.10.0
# matplotlib-inline==0.1.7
mistune==3.1.1
more-itertools==10.6.0
mpi4py==4.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy #==2.1.3
openmpi==0.0.0
openpyxl==3.1.5
packaging==24.1
pandas #==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
playwright==1.50.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==6.30.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyalex==0.15.1
pyarrow==19.0.0
pycparser==3.0
pydantic==2.11.4
pydantic_core==2.33.2
# pyee==12.1.2
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
psycopg[binary]==3.2.13
pytz==2025.1
pyusb==1.2.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
safetensors==0.5.3
# scikit-fuzzy==0.5.0
# scikit-learn #==1.5.2
# scipy #==1.14.1
# seaborn==0.13.2
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==75.1.0
six==1.16.0
# sklearn-compat==0.1.3
SLPP==1.2.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.5
stack-data==0.6.3
statsmodels==0.14.5
swifter==1.4.0
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.2
toolz==1.0.0
# torch==2.5.1
# torchaudio==2.5.1
# torchfile==0.1.0
# torchvision==0.20.1
# tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.53.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.26.5
wandb==0.19.10
wcwidth==0.2.13
webencodings==0.5.1
xlrd==2.0.1
xxhash==3.5.0
yarl==1.18.3
